+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Sequence and IteratorProtocol implementations to `Scru128Generator` to make it
+  work as infinite iterator
+
 ## v2.1.2 - 2022-06-11
 
 ### Fixed